changed type from int to short

+++ src/org/jgroups/protocols/PingHeader.java
@@ -1,4 +1,4 @@
-// $Id: PingHeader.java,v 1.6 2004/10/04 20:43:31 belaban Exp $
+// $Id: PingHeader.java,v 1.7 2004/10/08 13:15:46 belaban Exp $
 
 package org.jgroups.protocols;
 
@@ -9,16 +9,16 @@ import java.io.*;
 
 
 public class PingHeader extends Header implements Streamable {
-    public static final int GET_MBRS_REQ=1;   // arg = null
-    public static final int GET_MBRS_RSP=2;   // arg = PingRsp(local_addr, coord_addr)
+    public static final short GET_MBRS_REQ=1;   // arg = null
+    public static final short GET_MBRS_RSP=2;   // arg = PingRsp(local_addr, coord_addr)
 
-    public int type=0;
+    public short type=0;
     public PingRsp arg=null;
 
     public PingHeader() {
     } // for externalization
 
-    public PingHeader(int type, PingRsp arg) {
+    public PingHeader(short type, PingRsp arg) {
         this.type=type;
         this.arg=arg;
     }
@@ -44,18 +44,18 @@ public class PingHeader extends Header implements Streamable {
 
 
     public void writeExternal(ObjectOutput out) throws IOException {
-        out.writeInt(type);
+        out.writeShort(type);
         out.writeObject(arg);
     }
 
 
     public void readExternal(ObjectInput in) throws IOException, ClassNotFoundException {
-        type=in.readInt();
+        type=in.readShort();
         arg=(PingRsp)in.readObject();
     }
 
     public void writeTo(DataOutputStream outstream) throws IOException {
-        outstream.writeInt(type);
+        outstream.writeShort(type);
         if(arg != null) {
             outstream.write(1);
             arg.writeTo(outstream);
@@ -66,7 +66,7 @@ public class PingHeader extends Header implements Streamable {
     }
 
     public void readFrom(DataInputStream instream) throws IOException, IllegalAccessException, InstantiationException {
-        type=instream.readInt();
+        type=instream.readShort();
         int b=instream.read();
         if(b == 1) {
             arg=new PingRsp();
